move servser/database -> server/
