+++ server/routes/books.js
@@ -1,7 +1,7 @@
 const express = require('express');
 const router = new express.Router();
 const {check, validationResult} = require('express-validator');
-const models = require('../database/models');
+const models = require('../models');
 
 router.use(express.json());
 router.use(express.urlencoded({extended: true}));
